// src/app/core/enums/servicos.ts
export enum Servicos {
    faltaDeEnergia = 'falta-de-energia',
    religacao = 'religacao',
    historicoDeconsumo = 'historico-de-consumo',
    autoleitura = 'autoleitura',
    debitoAutomatico = 'debito-automatico',
    dataCerta = 'data-certa',
    lgpd = 'lgpd',
    minhaConta = 'minha-conta',
    segundaVia = 'consultar-debitos',
    alterarDataDeVencimento = 'data-certa',
    acompanheSeuPedido = 'acompanhe-seu-pedido',
    escolhaTipoFatura = 'escolha-tipo-fatura',
    faturaImpressa = 'fatura-impressa',
    faturaDigital = 'fatura-digital',
    faturaBraile = 'fatura-em-braile',
    entendaSuaConta = 'entenda-sua-conta',
    trocaDeTitularidade = 'troca-de-titularidade',
    memoriaDeMassa = 'memoria-de-massa',
    solicitacaoEnviada = 'solicitacao-enviada',
    segundaViaLogin = 'segunda-via-de-pagamento',
    cadastro = 'cadastro',
    aviso = 'aviso',
    multiloginAcesso = 'multilogin-acesso',
    multiloginCadastroCredenciado = 'multilogin-cadastro-credenciado',
    multiloginCadastroImobiliaria = 'multilogin-cadastro-imobiliario',
    multiloginCadastro = 'multilogin-cadastro',
    recuperarSenha = 'recuperar-senha',
    projetoParticular = 'projeto-particular',
    projetoParticularSolicitarAnalise = 'solicitar-analise',
    projetoParticularFormulario = 'formulario',
    unidadesConsumidoras = 'meus-imoveis',
    compartilharAcesso = 'compartilhar-acesso',
    meusImoveis = 'meus-imoveis',
    ligacaoNova = 'ligacao-nova',
    selecionarPerfilAcesso = 'selecionar-perfil-de-acesso'
}