// src/app/core/models/falta-de-energia/sub-rotas-falta-de-energia.ts
export enum SubRotasFaltaDeEnergia {
    FaltaDeEnergia = "",
    VerificarDisjuntor = "verificar-disjuntor",
    DisjuntorFuncionando = "disjuntor-funcionando",
    Avisos = "avisos",
    ConfiraSeusDados = "confira-seus-dados",
    DadosContato = "dados-contato",
    IluminacaoPublica = "iluminacao-publica",
    Passos = "passos",
    Problema = "problema",
    ProblemaResolvido = "problema-resolvido",
    ExibirAviso = "exibir-aviso"
}


export enum BreadCrumbFaltaDeEnergia {
    FaltaDeEnergia = "Falta de Energia",
    VerificarDisjuntor = "Verificar Disjuntor",
    DisjuntorFuncionando = "Disjuntor Funcionando",
    Avisos = "Avisos",
    ConfiraSeusDados = "Confira Seus Dados",
    DadosContato = "Dados Contato",
    IluminacaoPublica = "Iluminação Pública",
    Passos = "Passos",
    Problema = "Problema",
    ProblemaResolvido = "Problema Resolvido",
}